+++ src/main/java/com/peaksoft/model/entity/enums/Role.java
@@ -1,14 +1,38 @@
 package com.peaksoft.model.entity.enums;
 
+import com.peaksoft.model.User;
+import lombok.*;
 import org.springframework.security.core.GrantedAuthority;
 
-public enum Role implements GrantedAuthority  {
+import javax.persistence.*;
+import java.util.List;
 
-    ADMIN,
-    USER;
+@Entity
+@Table(name = "roles")
+@Setter
+@Getter
+@AllArgsConstructor
+@NoArgsConstructor
+@ToString
+public class Role implements GrantedAuthority {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.SEQUENCE)
+    @SequenceGenerator(name = "role_gen",
+            sequenceName = "role_seq",
+            allocationSize = 1)
+    private Long id;
+    private String roleName;
+    @ManyToMany(targetEntity = User.class,
+            cascade = {CascadeType.MERGE,
+                    CascadeType.DETACH,
+                    CascadeType.PERSIST,
+                    CascadeType.REFRESH})
+    private List<User> users;
+
 
     @Override
     public String getAuthority() {
-        return name();
+        return roleName;
     }
 }
