--- src/main/java/com/peaksoft/spring_boot/entity/User.java
@@ -1,81 +1,0 @@
-package com.peaksoft.spring_boot.entity;
-
-import lombok.*;
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
-import org.springframework.security.core.userdetails.UserDetails;
-
-import javax.persistence.*;
-import java.time.LocalDate;
-import java.util.ArrayList;
-import java.util.Collection;
-import java.util.List;
-
-@Entity
-@Table(name = "users")
-@Getter@Setter
-@AllArgsConstructor
-@NoArgsConstructor
-@ToString
-public class User implements UserDetails {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.SEQUENCE)
-    @SequenceGenerator(name = "user_gen",
-    sequenceName = "user_seq",
-    allocationSize = 1)
-    private Long id;
-    private String name;
-    private String surname;
-    private String email;
-    private String password;
-    @CreatedDate
-    private LocalDate created;
-
-    @ManyToMany(targetEntity = Role.class,
-    cascade = {CascadeType.MERGE,
-    CascadeType.PERSIST,
-    CascadeType.REFRESH,
-    CascadeType.DETACH},
-    fetch = FetchType.EAGER)
-    @JoinTable(name = "users_roles",
-            joinColumns = {@JoinColumn(name = "user_id")},
-            inverseJoinColumns = {@JoinColumn(name = "role_id")})
-    private List<Role> roles;
-
-
-    @Override
-    public Collection<? extends GrantedAuthority> getAuthorities() {
-        List<SimpleGrantedAuthority> grantedAuthorities = new ArrayList<>();
-        for(Role role : roles){
-            grantedAuthorities.add(new SimpleGrantedAuthority(role.getRoleName()));
-        }
-        return grantedAuthorities;
-    }
-
-    @Override
-    public String getUsername() {
-        return email;
-    }
-
-    @Override
-    public boolean isAccountNonExpired() {
-        return true;
-    }
-
-    @Override
-    public boolean isAccountNonLocked() {
-        return true;
-    }
-
-    @Override
-    public boolean isCredentialsNonExpired() {
-        return true;
-    }
-
-    @Override
-    public boolean isEnabled() {
-        return true;
-    }
-}
