+++ src/main/java/com/peaksoft/spring_boot_jwt_token/controller/TestController.java
@@ -1,4 +1,4 @@
-package com.peaksoft.spring_boot.controller;
+package com.peaksoft.spring_boot_jwt_token.controller;
 
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.GetMapping;
