+++ src/main/java/com/peaksoft/spring_boot_jwt_token/security/SecurityConfig.java
@@ -1,14 +1,13 @@
-package com.peaksoft.spring_boot.security;
+package com.peaksoft.spring_boot_jwt_token.security;
 
-import com.peaksoft.spring_boot.security.jwt.JwTokenFilter;
-import com.peaksoft.spring_boot.service.UserServiceImpl;
+import com.peaksoft.spring_boot_jwt_token.security.jwt.JwTokenFilter;
+import com.peaksoft.spring_boot_jwt_token.service.UserServiceImpl;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
-import org.springframework.security.config.annotation.method.configuration.EnableGlobalMethodSecurity;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
 import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
