--- src/main/resources/application-dev.properties
@@ -1,10 +1,0 @@
-spring.datasource.driver-class-name=org.postgresql.Driver
-spring.datasource.url=jdbc:postgresql://localhost:5432/java
-spring.datasource.username=postgres
-spring.datasource.password=1227
-spring.jpa.show-sql=true
-spring.jpa.hibernate.ddl-auto=create
-spring.jpa.database-platform=org.hibernate.dialect.PostgreSQL10Dialect
-spring.main.allow-bean-definition-overriding=true
-
-server.port=5050
